BlurBackgroundSliderView.java
When images does not fit sliderLayout the fitCenterCrop scale show empty space of image view as white which does not look good. So I made a new slider view which takes the blurred drawable of the original image and set it as a background and now we do not need to set the scale as fitXY as this doesn't show complete image. 
Also this customize slider type will give users an idea that they can make their own slider view with custom layouts but including all  the views required by the BaseSliderView.